+++ modules/alloy/src/main/ts/ephox/alloy/dragging/mouseortouch/MouseOrTouchDraggingTypes.ts
@@ -1,0 +1,7 @@
+import { CommonDraggingConfigSpec, DraggingConfig } from '../common/DraggingTypes';
+
+export interface MouseOrTouchDraggingConfigSpec extends CommonDraggingConfigSpec {
+  mode: 'mouseOrTouch';
+}
+
+export interface MouseOrTouchDraggingConfig extends DraggingConfig { }