+++ modules/tinymce/src/plugins/paste/main/ts/core/SmartPaste.ts
@@ -73,8 +73,8 @@ const smartInsertContent = function (editor: Editor, html: string) {
   });
 };
 
-const insertContent = function (editor: Editor, html: string) {
-  if (Settings.isSmartPasteEnabled(editor) === false) {
+const insertContent = function (editor: Editor, html: string, pasteAsText: boolean) {
+ if (pasteAsText || Settings.isSmartPasteEnabled(editor) === false) {
     pasteHtml(editor, html);
   } else {
     smartInsertContent(editor, html);